backend-requirements.txt
python-dotenv==1.0.0
requests==2.31.0
httpx==0.26.0
asyncio==3.4.3
typing-extensions==4.9.0
urllib3<2.0.0  # To avoid SSL warnings
yfinance==0.2.36  # For stock market data
plotly==5.18.0  # For charting
pandas==2.2.0  # For data manipulation
numpy==1.26.4  # Required by pandas and plotly
vaderSentiment==3.3.2  # For sentiment analysis
pytrends==4.9.2  # For Google Trends data
matplotlib==3.8.3  # For plotting
seaborn==0.13.2  # For statistical visualizations